# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "公共API服务平台",
        "title": "VAPIV API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/bilibili/video": {
            "get": {
                "tags": [
                    "内容数据"
                ],
                "summary": "B站视频信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频BV号",
                        "name": "bvid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/crypto/decrypt": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "核心服务"
                ],
                "summary": "AES解密",
                "parameters": [
                    {
                        "description": "解密请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CryptoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/crypto/encrypt": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "核心服务"
                ],
                "summary": "AES加密",
                "parameters": [
                    {
                        "description": "加密请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CryptoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/ip": {
            "get": {
                "tags": [
                    "核心服务"
                ],
                "summary": "IP地址查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP地址",
                        "name": "ip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/qq/avatar": {
            "get": {
                "tags": [
                    "内容数据"
                ],
                "summary": "QQ头像获取",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QQ号",
                        "name": "qq",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "头像尺寸",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CryptoReq": {
            "type": "object",
            "required": [
                "key",
                "text"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}